+++ tests/integration/PaymentMethods/IdealPaymentMethodTest.php
@@ -86,6 +86,7 @@ class IdealPaymentMethodTest extends BasePaymentMethodTest
      *
      * @return string payment reference id for the iDeal authorize transaction
      * @group connectionTest
+     * @throws \Exception
      */
     public function testAuthorize()
     {
@@ -96,9 +97,10 @@ class IdealPaymentMethodTest extends BasePaymentMethodTest
         $this->paymentObject->authorize();
 
         /* prepare request and send it to payment api */
-        $request = $this->paymentObject->getRequest()->convertToArray();
+        $request = $this->paymentObject->getRequest()->toArray();
         /** @var Response $response */
-        list(, $response) = $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
+        list($result, $response) =
+            $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
 
         /* test if config parameters exists */
         $configBankCountry = array('NL' => 'Niederlande');
@@ -115,6 +117,8 @@ class IdealPaymentMethodTest extends BasePaymentMethodTest
 
         $this->assertTrue($response->isSuccess(), 'Transaction failed : ' . print_r($response, 1));
         $this->assertFalse($response->isError(), 'authorize failed : ' . print_r($response->getError(), 1));
+
+        $this->logDataToDebug($result);
 
         return (string)$response->getPaymentReferenceId();
     }