+++ tests/integration/PaymentMethods/InvoiceB2CSecuredPaymentMethodTest.php
@@ -97,6 +97,7 @@ class InvoiceB2CSecuredPaymentMethodTest extends BasePaymentMethodTest
      * @return string payment reference id for the invoice authorize transaction
      * @group connectionTest
      * @test
+     * @throws \Exception
      */
     public function authorize()
     {
@@ -122,6 +123,8 @@ class InvoiceB2CSecuredPaymentMethodTest extends BasePaymentMethodTest
             'authorize failed : ' . print_r($this->paymentObject->getResponse()->getError(), 1)
         );
 
+        $this->logDataToDebug();
+
         return $this->authorizeReference = (string)$this->paymentObject->getResponse()->getPaymentReferenceId();
     }
 
@@ -134,6 +137,7 @@ class InvoiceB2CSecuredPaymentMethodTest extends BasePaymentMethodTest
      * @depends authorize
      * @group connectionTest
      * @test
+     * @throws \Exception
      */
     public function finalize($referenceId)
     {
@@ -156,6 +160,8 @@ class InvoiceB2CSecuredPaymentMethodTest extends BasePaymentMethodTest
             'reversal failed : ' . print_r($this->paymentObject->getResponse()->getError(), 1)
         );
 
+        $this->logDataToDebug();
+
         return $referenceId;
     }
 
@@ -168,6 +174,7 @@ class InvoiceB2CSecuredPaymentMethodTest extends BasePaymentMethodTest
      * @depends authorize
      * @group connectionTest
      * @test
+     * @throws \Exception
      */
     public function reversal($referenceId)
     {
@@ -192,6 +199,8 @@ class InvoiceB2CSecuredPaymentMethodTest extends BasePaymentMethodTest
             'reversal failed : ' . print_r($this->paymentObject->getResponse()->getError(), 1)
         );
 
+        $this->logDataToDebug();
+
         return (string)$this->paymentObject->getResponse()->getPaymentReferenceId();
     }
 
@@ -200,10 +209,10 @@ class InvoiceB2CSecuredPaymentMethodTest extends BasePaymentMethodTest
      *
      * @param string $referenceId reference id of the invoice to refund
      *
-     * @return string payment reference id of the invoice refund transaction
      * @depends authorize
      * @test
      * @group connectionTest
+     * @throws \Heidelpay\PhpPaymentApi\Exceptions\UndefinedTransactionModeException
      */
     public function refund($referenceId = null)
     {
@@ -216,7 +225,8 @@ class InvoiceB2CSecuredPaymentMethodTest extends BasePaymentMethodTest
         $this->paymentObject->refund((string)$referenceId);
 
         $this->assertEquals('IV.RF', $this->paymentObject->getRequest()->getPayment()->getCode());
-        return true;
+
+        $this->logDataToDebug();
     }
 
     //</editor-fold>