+++ tests/integration/PaymentMethods/PayPalPaymentMethodTest.php
@@ -84,12 +84,76 @@ class PayPalPaymentMethodTest extends BasePaymentMethodTest
     }
 
     /**
+     * Test case for a PayPal registration
+     *
+     * @return string payment reference id for the PayPal registration transaction
+     * @group connectionTest
+     * @throws \Exception
+     * @test
+     */
+    public function registration()
+    {
+        $timestamp = $this->getMethod(__METHOD__) . ' ' . date('Y-m-d H:i:s');
+        $this->paymentObject->getRequest()->basketData($timestamp, 23.12, $this->currency, $this->secret);
+        $this->paymentObject->getRequest()->async('DE', 'https://dev.heidelpay.de');
+
+        $this->paymentObject->registration();
+
+        /* prepare request and send it to payment api */
+        $request = $this->paymentObject->getRequest()->toArray();
+
+        /** @var Response $response */
+        list($result, $response) =
+            $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
+
+        $this->assertTrue($response->isSuccess(), 'Transaction failed : ' . print_r($response, 1));
+        $this->assertFalse($response->isError(), 'Transaction failed : ' . print_r($response->getError(), 1));
+
+        $this->logDataToDebug($result);
+
+        return (string)$response->getPaymentReferenceId();
+    }
+
+    /**
+     * Test case for a PayPal reregistration
+     *
+     * @group connectionTest
+     * @param null $referenceId
+     * @throws \Exception
+     * @depends registration
+     * @test
+     */
+    public function reregistration($referenceId = null)
+    {
+        $timestamp = $this->getMethod(__METHOD__) . ' ' . date('Y-m-d H:i:s');
+        $this->paymentObject->getRequest()->basketData($timestamp, 23.12, $this->currency, $this->secret);
+        $this->paymentObject->getRequest()->async('DE', 'https://dev.heidelpay.de');
+
+        $this->paymentObject->reregistration($referenceId);
+
+        /* prepare request and send it to payment api */
+        $request = $this->paymentObject->getRequest()->toArray();
+
+        /** @var Response $response */
+        list($result, $response) =
+            $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
+
+        $this->assertTrue($response->isSuccess(), 'Transaction failed : ' . print_r($response, 1));
+        $this->assertFalse($response->isError(), 'Transaction failed : ' . print_r($response->getError(), 1));
+
+        $this->logDataToDebug($result);
+    }
+
+    /**
      * Test case for a single PayPal authorisation
      *
      * @return string payment reference id for the PayPal authorize transaction
      * @group connectionTest
+     * @depends registration
+     * @throws \Exception
+     * @test
      */
-    public function testAuthorize()
+    public function authorize()
     {
         $timestamp = $this->getMethod(__METHOD__) . ' ' . date('Y-m-d H:i:s');
         $this->paymentObject->getRequest()->basketData($timestamp, 23.12, $this->currency, $this->secret);
@@ -98,14 +162,16 @@ class PayPalPaymentMethodTest extends BasePaymentMethodTest
         $this->paymentObject->authorize();
 
         /* prepare request and send it to payment api */
-        $request = $this->paymentObject->getRequest()->convertToArray();
+        $request = $this->paymentObject->getRequest()->toArray();
+
         /** @var Response $response */
-        list(, $response) = $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
+        list($result, $response) =
+            $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
 
         $this->assertTrue($response->isSuccess(), 'Transaction failed : ' . print_r($response, 1));
         $this->assertFalse($response->isError(), 'authorize failed : ' . print_r($response->getError(), 1));
 
-        return (string)$response->getPaymentReferenceId();
+        $this->logDataToDebug($result);
     }
 
     /**
@@ -113,8 +179,11 @@ class PayPalPaymentMethodTest extends BasePaymentMethodTest
      *
      * @return string payment reference id for the PayPal authorize transaction
      * @group connectionTest
+     * @depends registration
+     * @throws \Exception
+     * @test
      */
-    public function testDebit()
+    public function debit()
     {
         $timestamp = $this->getMethod(__METHOD__) . ' ' . date('Y-m-d H:i:s');
         $this->paymentObject->getRequest()->basketData($timestamp, 23.12, $this->currency, $this->secret);
@@ -123,13 +192,15 @@ class PayPalPaymentMethodTest extends BasePaymentMethodTest
         $this->paymentObject->debit();
 
         /* prepare request and send it to payment api */
-        $request = $this->paymentObject->getRequest()->convertToArray();
+        $request = $this->paymentObject->getRequest()->toArray();
+
         /** @var Response $response */
-        list(, $response) = $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
+        list($result, $response) =
+            $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
 
         $this->assertTrue($response->isSuccess(), 'Transaction failed : ' . print_r($response, 1));
         $this->assertFalse($response->isError(), 'debit failed : ' . print_r($response->getError(), 1));
 
-        return (string)$response->getPaymentReferenceId();
+        $this->logDataToDebug($result);
     }
 }