+++ tests/integration/PaymentMethods/PostFinanceEFinancePaymentMethodTest.php
@@ -89,6 +89,7 @@ class PostFinanceEFinancePaymentMethodTest extends BasePaymentMethodTest
      *
      * @return string payment reference id for the PostFinanceEFinance authorize transaction
      * @group connectionTest
+     * @throws \Exception
      */
     public function testAuthorize()
     {
@@ -99,12 +100,15 @@ class PostFinanceEFinancePaymentMethodTest extends BasePaymentMethodTest
         $this->paymentObject->authorize();
 
         /* prepare request and send it to payment api */
-        $request = $this->paymentObject->getRequest()->convertToArray();
+        $request = $this->paymentObject->getRequest()->toArray();
         /** @var Response $response */
-        list(, $response) = $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
+        list($result, $response) =
+            $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
 
         $this->assertTrue($response->isSuccess(), 'Transaction failed : ' . print_r($response, 1));
         $this->assertFalse($response->isError(), 'authorize failed : ' . print_r($response->getError(), 1));
+
+        $this->logDataToDebug($result);
 
         return (string)$response->getPaymentReferenceId();
     }