+++ tests/integration/PaymentMethods/SofortPaymentMethodTest.php
@@ -89,6 +89,7 @@ class SofortPaymentMethodTest extends BasePaymentMethodTest
      * @return string payment reference id for the sofort authorize transaction
      * @group connectionTest
      * @test
+     * @throws \Exception
      */
     public function authorize()
     {
@@ -99,12 +100,15 @@ class SofortPaymentMethodTest extends BasePaymentMethodTest
         $this->paymentObject->authorize();
 
         /* prepare request and send it to payment api */
-        $request = $this->paymentObject->getRequest()->convertToArray();
+        $request = $this->paymentObject->getRequest()->toArray();
         /** @var Response $response */
-        list(, $response) = $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
+        list($result, $response) =
+            $this->paymentObject->getRequest()->send($this->paymentObject->getPaymentUrl(), $request);
 
         $this->assertTrue($response->isSuccess(), 'Transaction failed : ' . print_r($response, 1));
         $this->assertFalse($response->isError(), 'authorize failed : ' . print_r($response->getError(), 1));
+
+        $this->logDataToDebug($result);
 
         return (string)$response->getPaymentReferenceId();
     }
@@ -118,6 +122,7 @@ class SofortPaymentMethodTest extends BasePaymentMethodTest
      * @depends authorize
      * @test
      * @group connectionTest
+     * @throws \Heidelpay\PhpPaymentApi\Exceptions\UndefinedTransactionModeException
      */
     public function refund($referenceId = null)
     {
@@ -130,6 +135,7 @@ class SofortPaymentMethodTest extends BasePaymentMethodTest
         $this->paymentObject->refund((string)$referenceId);
 
         $this->assertEquals('OT.RF', $this->paymentObject->getRequest()->getPayment()->getCode());
-        return true;
+
+        $this->logDataToDebug();
     }
 }